LayoutTests/platform/ios/imported/w3c/web-platform-tests/digital-credentials/get.tentative.https-expected.txt



PASS Type conversion happens on the data member of the DigitalCredentialGetRequest object.
PASS Calling navigator.credentials.get() without a digital member same origin.
PASS navigator.credentials.get() API rejects if there are no credential request.
PASS navigator.credentials.get() API rejects if there are no credential request for same-origin iframe.
PASS navigator.credentials.get() API rejects if there are no credential request in cross-origin iframe.
PASS navigator.credentials.get() promise is rejected if called with an aborted controller.
PASS navigator.credentials.get() promise is rejected if called with an aborted controller in same-origin iframe.
PASS navigator.credentials.get() promise is rejected if called with an aborted signal in cross-origin iframe.
PASS navigator.credentials.get() promise is rejected if abort controller is aborted after call to get().
PASS navigator.credentials.get() promise is rejected if abort controller is aborted after call to get() in cross-origin iframe.
FAIL Mediation is implicitly required and hence ignored. Request is aborted regardless. promise_rejects_js: function "function() { throw e; }" threw object "InvalidStateError: A credential picker operation is already in progress." ("InvalidStateError") expected instance of function "function Error() {
    [native code]
}" ("Error")
PASS Throws TypeError when request data is not JSON stringifiable.
PASS `requests` field is required in the options object.
